pyproject.toml
[tool.poetry]
name = "code-expiry"
version = "0.1.0"
description = "A web application for sharing code snippets with automatic expiry"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
djangorestframework = "^3.14.0"
python-dateutil = "^2.8.2"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 